Split PairController into GroupPairController
Split GroupPairController off of PairController.
Move pair rendering things into SharedView.
Move pair fetching and generating into PairMaker lib.
Correct controller tests.
