+++ web/controllers/controller_helpers.ex
@@ -2,6 +2,7 @@ defmodule Pairmotron.ControllerHelpers do
 
   import Plug.Conn
   import Phoenix.Controller
+  alias Pairmotron.{Repo, PairRetro}
 
   @doc """
   Renders a 404 message.
@@ -51,5 +52,14 @@ defmodule Pairmotron.ControllerHelpers do
   """
   def is_admin?(user) do
     user.is_admin
+  end
+
+  @doc """
+  Assigns a retro to the current user for a given year and week
+  """
+  def assign_current_user_pair_retro_for_week(conn, year, week) do
+    current_user = conn.assigns[:current_user]
+    retro = Repo.one(PairRetro.retro_for_user_and_week(current_user, year, week))
+    Plug.Conn.assign(conn, :current_user_retro_for_week, retro)
   end
 end
