+++ web/templates/profile/show.html.eex
@@ -34,6 +34,11 @@
               </a>
             </td>
             <td>
+              <a href=<%= group_pair_path(@conn, :index, group)%>>
+                Pairs
+              </a>
+            </td>
+            <td>
               <%= if group.owner_id == @user.id do %>
                 <a href=<%= group_path(@conn, :edit, group)%>>Manage</a>
               <% end %>
